web_counter/requirements.txt
aiofiles==25.1.0
fastapi==0.124.4
httpx==0.28.1
uvicorn==0.38.0
PyYAML==6.0.3